ref: Improve Integration Tests a bit
- Creates a `utils` module that re-exports relevant things from `symbolicator_test`, and is itself re-exported from the root.
- Simplifies creating symbolication requests a bit.
- Moves tests for external public symbol servers to its dedicated file.
