+++ crates/symbolicator-service/tests/integration/e2e.rs
@@ -1,23 +1,24 @@
 use std::path::Path;
 use std::sync::Arc;
 
-use symbolicator_service::services::symbolication::{StacktraceOrigin, SymbolicateStacktraces};
-use symbolicator_service::types::{
-    FrameStatus, ObjectDownloadInfo, ObjectFileStatus, RawObjectInfo, RawStacktrace, Scope,
-};
+use symbolicator_service::services::symbolication::SymbolicateStacktraces;
+use symbolicator_service::types::{FrameStatus, ObjectDownloadInfo, ObjectFileStatus, Scope};
 use symbolicator_sources::{
     DirectoryLayoutType, FileType, FilesystemSourceConfig, HttpSourceConfig, RemoteFileUri,
     SentrySourceConfig, SourceConfig, SourceId,
 };
-use symbolicator_test::{assert_snapshot, fixture, source_config, Server};
 
-use crate::symbolication::{get_symbolication_request, setup_service};
+use crate::{
+    assert_snapshot, example_request, fixture, make_symbolication_request, setup_service,
+    source_config, Server,
+};
 
 // FIXME: Using this fixture in combination with the public microsoft symbol server means that
 // we are downloading multiple of megabytes on each test run, which we should ideally avoid by
 // testing with a local http source instead.
-fn request_fixture() -> (Vec<RawObjectInfo>, Vec<RawStacktrace>) {
-    let modules = serde_json::from_str(
+fn request_fixture(sources: Vec<SourceConfig>) -> SymbolicateStacktraces {
+    make_symbolication_request(
+        sources,
         r#"[{
           "type": "pe",
           "debug_id": "ff9f9f78-41db-88f0-cded-a9e1e9bff3b5-1",
@@ -26,16 +27,11 @@ fn request_fixture() -> (Vec<RawObjectInfo>, Vec<RawStacktrace>) {
           "image_addr": "0x749d0000",
           "image_size": 851968
         }]"#,
-    )
-    .unwrap();
-    let stacktraces = serde_json::from_str(
         r#"[{
           "registers": {"eip": "0x0000000001509530"},
           "frames": [{"instruction_addr": "0x749e8630"}]
         }]"#,
     )
-    .unwrap();
-    (modules, stacktraces)
 }
 
 /// tests that nothing happens when no source is supplied
@@ -43,17 +39,7 @@ fn request_fixture() -> (Vec<RawObjectInfo>, Vec<RawStacktrace>) {
 async fn test_no_sources() {
     let (symbolication, cache_dir) = setup_service(|_| ());
 
-    let (modules, stacktraces) = request_fixture();
-
-    let request = SymbolicateStacktraces {
-        modules: modules.into_iter().map(From::from).collect(),
-        stacktraces,
-        signal: None,
-        origin: StacktraceOrigin::Symbolicate,
-        sources: Arc::new([]),
-        scope: Default::default(),
-    };
-
+    let request = request_fixture(vec![]);
     let response = symbolication.symbolicate(request).await;
 
     assert_snapshot!(response.unwrap());
@@ -66,21 +52,12 @@ async fn test_no_sources() {
 #[tokio::test]
 async fn test_sources_filetypes() {
     let (symbolication, _cache_dir) = setup_service(|_| ());
-
-    let (modules, stacktraces) = request_fixture();
-
     // This symbol source is filtering for only `mach_code` files.
     let hitcounter = Server::new();
 
-    let request = SymbolicateStacktraces {
-        modules: modules.into_iter().map(From::from).collect(),
-        stacktraces,
-        signal: None,
-        origin: StacktraceOrigin::Symbolicate,
-        sources: Arc::new([hitcounter.source("not-found", "/respond_statuscode/404/")]),
-        scope: Default::default(),
-    };
-
+    let request = request_fixture(vec![
+        hitcounter.source("not-found", "/respond_statuscode/404/")
+    ]);
     let response = symbolication.symbolicate(request).await;
 
     assert_eq!(hitcounter.accesses(), 0);
@@ -88,12 +65,10 @@ async fn test_sources_filetypes() {
     assert_snapshot!(response.unwrap());
 }
 
-/// tests that sourcce `path_patterns` work as expected
+/// tests that source `path_patterns` work as expected
 #[tokio::test]
 async fn test_path_patterns() {
     let (symbolication, _cache_dir) = setup_service(|_| ());
-
-    let (modules, stacktraces) = request_fixture();
 
     let patterns = [
         (Some("\"?:/windows/**\""), true),
@@ -118,15 +93,7 @@ async fn test_path_patterns() {
             files,
         }));
 
-        let request = SymbolicateStacktraces {
-            modules: modules.iter().cloned().map(From::from).collect(),
-            stacktraces: stacktraces.clone(),
-            signal: None,
-            origin: StacktraceOrigin::Symbolicate,
-            sources: Arc::new([source]),
-            scope: Default::default(),
-        };
-
+        let request = request_fixture(vec![source]);
         let mut response = symbolication.symbolicate(request).await.unwrap();
 
         let mut module = response.modules.pop().unwrap();
@@ -150,7 +117,6 @@ async fn test_path_patterns() {
 #[tokio::test]
 async fn test_no_permission() {
     let (symbolication, _cache_dir) = setup_service(|_| ());
-
     let hitcounter = Server::new();
 
     let sources = vec![
@@ -180,7 +146,7 @@ async fn test_no_permission() {
         .unwrap(),
     ];
 
-    let request = get_symbolication_request(sources);
+    let request = example_request(sources);
     let mut response = symbolication.symbolicate(request).await.unwrap();
     let candidates = response.modules.pop().unwrap().candidates.0;
 
@@ -224,6 +190,7 @@ async fn test_no_permission() {
 /// to `127.0.0.1`.
 #[tokio::test]
 async fn test_reserved_ip_addresses() {
+    let (symbolication, _cache_dir) = setup_service(|_| ());
     let hitcounter = Server::new();
 
     let files = source_config(DirectoryLayoutType::Native, vec![FileType::MachCode]);
@@ -255,9 +222,7 @@ async fn test_reserved_ip_addresses() {
         files,
     })));
 
-    let request = get_symbolication_request(sources);
-
-    let (symbolication, _cache_dir) = setup_service(|_| ());
+    let request = example_request(sources);
     let mut response = symbolication.symbolicate(request.clone()).await.unwrap();
     let candidates = response.modules.pop().unwrap().candidates.0;
 
@@ -329,22 +294,12 @@ async fn test_reserved_ip_addresses() {
 #[tokio::test]
 async fn test_redirects() {
     let (symbolication, _cache_dir) = setup_service(|_| ());
-
-    let (modules, stacktraces) = request_fixture();
-
     let hitcounter = Server::new();
+
     let config = source_config(DirectoryLayoutType::Symstore, vec![FileType::Pdb]);
     let source = hitcounter.source_with_config("hitcounter", "redirect/msdl", config);
 
-    let request = SymbolicateStacktraces {
-        modules: modules.iter().cloned().map(From::from).collect(),
-        stacktraces: stacktraces.clone(),
-        signal: None,
-        origin: StacktraceOrigin::Symbolicate,
-        sources: Arc::new([source]),
-        scope: Default::default(),
-    };
-
+    let request = request_fixture(vec![source]);
     let mut response = symbolication.symbolicate(request).await.unwrap();
 
     let mut module = response.modules.pop().unwrap();
@@ -359,9 +314,6 @@ async fn test_redirects() {
 #[tokio::test]
 async fn test_unreachable_bucket() {
     let (symbolication, _cache_dir) = setup_service(|_| ());
-
-    let (modules, stacktraces) = request_fixture();
-
     let hitcounter = Server::new();
 
     let tys = ["http", "sentry"];
@@ -384,15 +336,7 @@ async fn test_unreachable_bucket() {
                 }))
             };
 
-            let request = SymbolicateStacktraces {
-                modules: modules.iter().cloned().map(From::from).collect(),
-                stacktraces: stacktraces.clone(),
-                signal: None,
-                origin: StacktraceOrigin::Symbolicate,
-                sources: Arc::new([source]),
-                scope: Default::default(),
-            };
-
+            let request = request_fixture(vec![source]);
             let mut response = symbolication.symbolicate(request).await.unwrap();
 
             let mut module = response.modules.pop().unwrap();
@@ -423,9 +367,6 @@ async fn test_unreachable_bucket() {
 #[tokio::test]
 async fn test_lookup_deduplication() {
     let (symbolication, _cache_dir) = setup_service(|_| ());
-
-    let (modules, stacktraces) = request_fixture();
-
     let hitcounter = Server::new();
 
     let mut config = source_config(DirectoryLayoutType::Symstore, vec![FileType::Pdb]);
@@ -434,24 +375,12 @@ async fn test_lookup_deduplication() {
         config.is_public = is_public;
         let source =
             hitcounter.source_with_config(&format!("test-{is_public}"), "msdl/", config.clone());
+        let request = request_fixture(vec![source]);
 
         let requests = (0..20).map(|_| {
-            let modules = modules.iter().cloned().map(From::from).collect();
-            let stacktraces = stacktraces.clone();
             let symbolication = symbolication.clone();
-            let source = source.clone();
-            async move {
-                let request = SymbolicateStacktraces {
-                    modules,
-                    stacktraces,
-                    signal: None,
-                    origin: StacktraceOrigin::Symbolicate,
-                    sources: Arc::new([source]),
-                    scope: Default::default(),
-                };
-
-                symbolication.symbolicate(request).await.unwrap()
-            }
+            let request = request.clone();
+            async move { symbolication.symbolicate(request).await.unwrap() }
         });
 
         let responses = futures::future::join_all(requests).await;
@@ -496,8 +425,6 @@ fn get_files(path: impl AsRef<Path>) -> Vec<(String, u64)> {
 /// Tests caching side-effects, like cache files written and hits to the symbol source.
 #[tokio::test]
 async fn test_basic_windows() {
-    let (modules, stacktraces) = request_fixture();
-
     let hitcounter = Server::new();
 
     let mut config = source_config(DirectoryLayoutType::Symstore, vec![FileType::Pdb]);
@@ -513,14 +440,8 @@ async fn test_basic_windows() {
                 }
             });
 
-            let request = SymbolicateStacktraces {
-                modules: modules.iter().cloned().map(From::from).collect(),
-                stacktraces: stacktraces.clone(),
-                signal: None,
-                origin: StacktraceOrigin::Symbolicate,
-                sources: Arc::new([source]),
-                scope: Scope::Scoped("myscope".into()),
-            };
+            let mut request = request_fixture(vec![source]);
+            request.scope = Scope::Scoped("myscope".into());
 
             for i in 0..2 {
                 let response = symbolication.symbolicate(request.clone()).await.unwrap();
