+++ crates/symbolicator-service/tests/integration/main.rs
@@ -2,5 +2,9 @@
 
 pub mod e2e;
 pub mod process_minidump;
+pub mod public_sources;
 pub mod source_errors;
 pub mod symbolication;
+pub mod utils;
+
+pub use utils::*;
